server2/pyproject.toml
[project]
name = "server2"
version = "0.1.0"
description = "Amchich backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "async-lru>=2.0.5",
    "cryptography>=45.0.5",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.13",
    "types-pyjwt>=1.7.1",
]
[tool.ruff]
line-length = 95

[tool.ruff.lint]
select = [
    "AIR",
    "ERA",
    "FAST",
    "YTT",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FIX",
    "FA",
    "INT",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "SLOT",
    "TID",
    "TC",
    "ARG",
    "PTH",
    "FLY",
    "I",
    "C90",
    "NPY",
    "PD",
    "N",
    "PERF",
    "E",
    "W",
    "F",
    "PGH",
    "PL",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "UP",
    "FURB",
    "RUF",
    "TRY",
]
ignore = ["COM812"]


[tool.mypy]
plugins = ["pydantic.mypy"]
